+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Changelog
 ### Fixes
 
 - **[IR]**: Fix `IllegalArgumentException` thrown when there are multiple top-level functions with the same name but only one is annotated with `@Inject`.
+- **[IR]**: Only store a given binding container's own provider factories in metro metadata. This resolves a bug where we could end up duplicate-processing upstream providers in dynamic factories.
 
 ### Contributors